+++ lib/bibliothecary/analyser/analysis.rb
@@ -40,10 +40,10 @@ module Bibliothecary
         # If your Parser needs to return multiple responses for one file, please override this method
         # For example see conda.rb
         kind = determine_kind_from_info(info)
-        dependencies_result = parse_file(info.relative_path, info.contents, options: options)
-        dependencies_result = DependenciesResult.new if dependencies_result.nil? # work around any legacy parsers that return nil
+        parser_result = parse_file(info.relative_path, info.contents, options: options)
+        parser_result = ParserResult.new if parser_result.nil? # work around any legacy parsers that return nil
 
-        Bibliothecary::Analyser.create_analysis(platform_name, info.relative_path, kind, dependencies_result)
+        Bibliothecary::Analyser.create_analysis(platform_name, info.relative_path, kind, parser_result)
       rescue Bibliothecary::FileParsingError => e
         Bibliothecary::Analyser.create_error_analysis(platform_name, info.relative_path, kind, e.message, e.location)
       end