+++ lib/bibliothecary/multi_parsers/dependencies_csv.rb
@@ -148,7 +148,7 @@ module Bibliothecary
             )
           end
 
-        DependenciesResult.new(dependencies: dependencies)
+        ParserResult.new(dependencies: dependencies)
       end
     end
   end