+++ lib/bibliothecary/multi_parsers/json_runtime.rb
@@ -15,7 +15,7 @@ module Bibliothecary
           )
         end
 
-        DependenciesResult.new(dependencies: dependencies)
+        ParserResult.new(dependencies: dependencies)
       end
     end
   end