+++ lib/bibliothecary/parser_result.rb
@@ -1,11 +1,11 @@
 # frozen_string_literal: true
 
 module Bibliothecary
-  # DependenciesResult bundles together a list of dependencies and the project name.
+  # ParserResult bundles together a list of dependencies and the project name.
   #
   # @attr_reader [Array<Dependency>] dependencies The list of Dependency objects
   # @attr_reader [String,nil] project_name The name of the project
-  class DependenciesResult
+  class ParserResult
     FIELDS = %i[
       dependencies
       project_name
@@ -14,7 +14,7 @@ module Bibliothecary
     attr_reader(*FIELDS)
 
     def initialize(
-      dependencies: [],
+      dependencies:,
       project_name: nil
     )
       @dependencies = dependencies
@@ -25,6 +25,10 @@ module Bibliothecary
       FIELDS.all? { |f| public_send(f) == other.public_send(f) }
     end
     alias == eql?
+
+    def to_h
+      FIELDS.to_h { |f| [f, public_send(f)] }
+    end
 
     def hash
       FIELDS.map { |f| public_send(f) }.hash