+++ lib/bibliothecary/parsers/bower.rb
@@ -22,7 +22,7 @@ module Bibliothecary
         json = JSON.parse(file_contents)
         dependencies = map_dependencies(json, "dependencies", "runtime", options.fetch(:filename, nil)) +
                        map_dependencies(json, "devDependencies", "development", options.fetch(:filename, nil))
-        DependenciesResult.new(dependencies: dependencies)
+        ParserResult.new(dependencies: dependencies)
       end
     end
   end