+++ lib/bibliothecary/parsers/cpan.rb
@@ -28,13 +28,13 @@ module Bibliothecary
         dependencies = manifest["prereqs"].map do |_group, deps|
           map_dependencies(deps, "requires", "runtime", options.fetch(:filename, nil))
         end.flatten
-        DependenciesResult.new(dependencies: dependencies)
+        ParserResult.new(dependencies: dependencies)
       end
 
       def self.parse_yaml_manifest(file_contents, options: {})
         manifest = YAML.load file_contents
         dependencies = map_dependencies(manifest, "requires", "runtime", options.fetch(:filename, nil))
-        DependenciesResult.new(dependencies: dependencies)
+        ParserResult.new(dependencies: dependencies)
       end
     end
   end