+++ lib/bibliothecary/parsers/nuget.rb
@@ -62,7 +62,7 @@ module Bibliothecary
             platform: platform_name
           )
         end
-        DependenciesResult.new(dependencies: dependencies)
+        ParserResult.new(dependencies: dependencies)
       end
 
       def self.parse_packages_lock_json(file_contents, options: {})
@@ -91,9 +91,9 @@ module Bibliothecary
 
           # Note, frameworks can be empty, so remove empty ones and then return the last sorted item if any
           frameworks.delete_if { |_k, v| v.empty? }
-          return DependenciesResult.new(dependencies: frameworks[frameworks.keys.max]) unless frameworks.empty?
+          return ParserResult.new(dependencies: frameworks[frameworks.keys.max]) unless frameworks.empty?
         end
-        DependenciesResult.new(dependencies: [])
+        ParserResult.new(dependencies: [])
       end
 
       def self.parse_packages_config(file_contents, options: {})
@@ -107,9 +107,9 @@ module Bibliothecary
             platform: platform_name
           )
         end
-        DependenciesResult.new(dependencies: dependencies)
+        ParserResult.new(dependencies: dependencies)
       rescue StandardError
-        DependenciesResult.new(dependencies: [])
+        ParserResult.new(dependencies: [])
       end
 
       def self.parse_csproj(file_contents, options: {})
@@ -166,9 +166,9 @@ module Bibliothecary
           end
 
         dependencies = packages.uniq(&:name)
-        DependenciesResult.new(dependencies: dependencies)
+        ParserResult.new(dependencies: dependencies)
       rescue StandardError
-        DependenciesResult.new(dependencies: [])
+        ParserResult.new(dependencies: [])
       end
 
       def self.parse_nuspec(file_contents, options: {})
@@ -182,9 +182,9 @@ module Bibliothecary
             platform: platform_name
           )
         end
-        DependenciesResult.new(dependencies: dependencies)
+        ParserResult.new(dependencies: dependencies)
       rescue StandardError
-        DependenciesResult.new(dependencies: [])
+        ParserResult.new(dependencies: [])
       end
 
       def self.parse_paket_lock(file_contents, options: {})
@@ -201,7 +201,7 @@ module Bibliothecary
         end
         # we only have to enforce uniqueness by name because paket ensures that there is only the single version globally in the project
         dependencies = packages.uniq(&:name)
-        DependenciesResult.new(dependencies: dependencies)
+        ParserResult.new(dependencies: dependencies)
       end
 
       def self.parse_project_assets_json(file_contents, options: {})
@@ -229,9 +229,9 @@ module Bibliothecary
 
           # Note, frameworks can be empty, so remove empty ones and then return the last sorted item if any
           frameworks.delete_if { |_k, v| v.empty? }
-          return DependenciesResult.new(dependencies: frameworks[frameworks.keys.max]) unless frameworks.empty?
+          return ParserResult.new(dependencies: frameworks[frameworks.keys.max]) unless frameworks.empty?
         end
-        DependenciesResult.new(dependencies: [])
+        ParserResult.new(dependencies: [])
       end
     end
   end