+++ lib/bibliothecary/parsers/shard.rb
@@ -25,14 +25,14 @@ module Bibliothecary
       def self.parse_yaml_lockfile(file_contents, options: {})
         manifest = YAML.load file_contents
         dependencies = map_dependencies(manifest, "shards", "runtime", options.fetch(:filename, nil))
-        Bibliothecary::DependenciesResult.new(dependencies: dependencies)
+        Bibliothecary::ParserResult.new(dependencies: dependencies)
       end
 
       def self.parse_yaml_manifest(file_contents, options: {})
         manifest = YAML.load file_contents
         dependencies = map_dependencies(manifest, "dependencies", "runtime", options.fetch(:filename, nil)) +
                        map_dependencies(manifest, "development_dependencies", "runtime", options.fetch(:filename, nil))
-        Bibliothecary::DependenciesResult.new(dependencies: dependencies)
+        Bibliothecary::ParserResult.new(dependencies: dependencies)
       end
 
       def self.map_dependencies(hash, key, type, source = nil)